requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
gcloud==0.18.3
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.1
oauth2client==4.1.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==2.0.1
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
gunicorn